+++ Customer/Customer.Repository/Repositories/CustomerRepository.cs
@@ -11,46 +11,7 @@ namespace Customer.Repository.Repositories
     {
         public void CreateCustomer(string name)
         {
-            
+            throw new NotImplementedException();
         }
-
-
-        // ---- 以下のメソッドは練習用のサンプルとして作ったもので、本来は上のCreateCustomerを都度修正していくことになります ----
-        public bool CreateCustomer01(string name)
-        {
-            return true;
-        }
-
-        public bool CreateCustomer02(string name)
-        {
-            return true;
-        }
-
-        public bool CreateCustomer02_modified(string name)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-                return false;
-
-
-            return true;
-
-        }
-
-        public bool CreateCustomer02_modified02(string name)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-                return false;
-
-            char[] forbiddenChars = { '!', '@', '#', '$', '%', '&', '*', '\\' };
-            if (name.Any(x => forbiddenChars.Contains(x)))
-                return false;
-
-            return true;
-        }
-
-
-
-
-
     }
 }
